requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
email-validator==2.1.0.post1
garminconnect==0.1.50
requests==2.26.0
pymongo==3.12.0
gunicorn==20.1.0

# Machine Learning Dependencies
tensorflow==2.8.0
scikit-learn==1.0.2
numpy==1.21.4
pandas==1.3.5
joblib==1.1.0
matplotlib==3.5.1
seaborn==0.11.2

# API Documentation
flask-swagger-ui==3.36.0 